+++ prettier.config.mjs
@@ -1,0 +1,9 @@
+const prettierConfig = {
+  printWidth: 100,
+  semi: true,
+  singleQuote: false,
+  tabWidth: 2,
+  trailingComma: "all",
+};
+
+export default prettierConfig;